Adding argparse. Writing attempt log to file
Argparse added so that I can display command line args.

Writing the "attempt log" to file. At the moment defaults to "sort.log"
but will use argparse to allow user to select where it goes. The attempt
log will display the list of groups as well as the "final submission
date" so that any late penalties can be applied.