Confirmation if destination directory exists
Confirms with the user if they want to sort into destination directory
if the directory is already there (does not check if the directory is
empty though)

For each attempt, the attempt folder is first deleted first if any
exists to clean up any files inside. This means that there may be extra
folders lying about with old attempts.