Argument validation. More changes to destination handling
Arguments are now validated.

Destination cannot be the same as the Working Directory.
Destination cannot be an existing file.

If Destination is an existing folder, it is deleted and recreated
(all its contents destroyed) after confirmation from user.
Then all the copying happens. This is to make sure the
destination directory is always "clean".